requirements.txt
altair==6.0.0
attrs==26.1.0
blinker==1.9.0
cachetools==7.0.5
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.3.2
colorama==0.4.6
cryptography==46.0.7
Faker==40.12.0
gitdb==4.0.12
GitPython==3.1.46
greenlet==3.3.2
idna==3.11
Jinja2==3.1.6
joblib==1.5.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
loguru==0.7.3
MarkupSafe==3.0.3
narwhals==2.19.0
numpy==2.2.6
packaging==26.0
pandas==2.3.3
pillow==12.2.0
plotly==6.7.0
protobuf==7.34.1
pyarrow==23.0.1
pycparser==3.0
pydeck==0.9.1
PyMySQL==1.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pytz==2026.1.post1
referencing==0.37.0
requests==2.33.1
rpds-py==0.30.0
scikit-learn==1.7.2
scipy==1.15.3
six==1.17.0
smmap==5.0.3
SQLAlchemy==2.0.49
streamlit==1.56.0
tenacity==9.1.4
threadpoolctl==3.6.0
toml==0.10.2
tornado==6.5.5
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
watchdog==6.0.0
win32_setctime==1.2.0
